requirements.txt
# CryptoWatch - 依赖包列表
# Python 3.9+

# ===== 核心依赖 =====
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0

# ===== LLM框架 =====
langgraph>=0.0.20
langchain>=0.1.0

# ===== 数据处理 =====
pandas>=2.0.0

# ===== 开发工具 =====
pytest>=7.4.0
black>=23.0.0
